# docs/swagger.yaml
definitions:
  models.UserChart:
    properties:
      freeCount:
        type: integer
      id:
        type: integer
      token:
        type: integer
      userAccount:
        type: string
      userAvatar:
        type: string
      userId:
        type: integer
    type: object
  r.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  requests.ChartQueryRequest:
    properties:
      chartType:
        type: string
      goal:
        type: string
      key:
        type: string
      pageNum:
        type: integer
      pageSize:
        type: integer
      userId:
        type: integer
    type: object
  requests.LoginRequest:
    properties:
      userAccount:
        type: string
      userPassword:
        type: string
    type: object
  requests.Page:
    properties:
      pageNum:
        type: integer
      pageSize:
        type: integer
    type: object
  requests.RegisterRequest:
    properties:
      checkPassword:
        type: string
      userAccount:
        type: string
      userPassword:
        type: string
    type: object
  response.BiResp:
    properties:
      genChart:
        type: string
      genResult:
        type: string
    type: object
  serializers.CurrentUser:
    properties:
      id:
        description: 主键ID
        type: integer
      userAccount:
        type: string
      userAvatar:
        type: string
      userName:
        type: string
      userRole:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BI Pro API
  version: "1.0"
paths:
  /chart/all_list:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 查询请求参数
        in: body
        name: ChartQueryRequest
        required: true
        schema:
          $ref: '#/definitions/requests.ChartQueryRequest'
      produces:
      - application/json
      responses:
        "0":
          description: 成功
          schema:
            $ref: '#/definitions/response.BiResp'
        "40002":
          description: 参数错误
          schema:
            $ref: '#/definitions/r.Response'
        "40003":
          description: 系统错误
          schema:
            $ref: '#/definitions/r.Response'
      summary: ListAllChart
      tags:
      - ChartApi
  /chart/gen:
    post:
      consumes:
      - multipart/form-data
      description: 通过上传文件和发送请求生成图表
      parameters:
      - description: 要上传的文件
        in: formData
        name: file
        required: true
        type: file
      - description: 生成图表的目标
        in: formData
        name: goal
        required: true
        type: string
      - description: 图表类型
        in: formData
        name: chartType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.BiResp'
        "40002":
          description: 参数错误
          schema:
            $ref: '#/definitions/r.Response'
        "40003":
          description: 系统错误
          schema:
            $ref: '#/definitions/r.Response'
      summary: 生成图表
      tags:
      - ChartApi
  /chart/list:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 查询请求参数
        in: body
        name: ChartQueryRequest
        required: true
        schema:
          $ref: '#/definitions/requests.ChartQueryRequest'
      produces:
      - application/json
      responses:
        "0":
          description: 成功
          schema:
            $ref: '#/definitions/response.BiResp'
        "40002":
          description: 参数错误
          schema:
            $ref: '#/definitions/r.Response'
        "40003":
          description: 系统错误
          schema:
            $ref: '#/definitions/r.Response'
      summary: Chart List
      tags:
      - ChartApi
  /table/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录请求参数
        in: body
        name: pageRequest
        required: true
        schema:
          $ref: '#/definitions/requests.Page'
      produces:
      - application/json
      responses:
        "0":
          description: 成功
          schema:
            items:
              $ref: '#/definitions/models.UserChart'
            type: array
        "40002":
          description: 参数错误
          schema:
            $ref: '#/definitions/r.Response'
        "40003":
          description: 系统错误
          schema:
            $ref: '#/definitions/r.Response'
      summary: ListUserTable
      tags:
      - TableApi
  /user/current:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "0":
          description: 成功
          schema:
            $ref: '#/definitions/serializers.CurrentUser'
        "40005":
          description: 获取当前用户信息失败
          schema:
            $ref: '#/definitions/r.Response'
      summary: Current
      tags:
      - UserApi
  /user/list:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "0":
          description: 成功
          schema:
            $ref: '#/definitions/r.Response'
        "40002":
          description: 参数错误
          schema:
            $ref: '#/definitions/r.Response'
        "40003":
          description: 系统错误
          schema:
            $ref: '#/definitions/r.Response'
      summary: List
      tags:
      - UserApi
  /user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录请求参数
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/requests.LoginRequest'
      produces:
      - application/json
      responses:
        "0":
          description: 成功
          schema:
            $ref: '#/definitions/r.Response'
        "40002":
          description: 参数错误
          schema:
            $ref: '#/definitions/r.Response'
        "40003":
          description: 系统错误
          schema:
            $ref: '#/definitions/r.Response'
      summary: User Login
      tags:
      - UserApi
  /user/logout:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "0":
          description: 成功
          schema:
            $ref: '#/definitions/r.Response'
        "40002":
          description: 参数错误
          schema:
            $ref: '#/definitions/r.Response'
        "40003":
          description: 系统错误
          schema:
            $ref: '#/definitions/r.Response'
      summary: Logout
      tags:
      - UserApi
  /user/refresh_token:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "0":
          description: 成功
          schema:
            $ref: '#/definitions/r.Response'
        "40005":
          description: 认证失败
          schema:
            $ref: '#/definitions/r.Response'
      summary: RefreshToken
      tags:
      - UserApi
  /user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册请求参数
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/requests.RegisterRequest'
      produces:
      - application/json
      responses:
        "0":
          description: 成功
          schema:
            $ref: '#/definitions/r.Response'
        "40002":
          description: 参数错误
          schema:
            $ref: '#/definitions/r.Response'
        "40003":
          description: 系统错误
          schema:
            $ref: '#/definitions/r.Response'
      summary: User Register
      tags:
      - UserApi
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"